requirements.txt
Flask==2.0.1
Flask-Login==0.5.0
Flask-WTF==0.15.1
Flask-Session==0.4.0
Werkzeug==2.0.1
pymongo[srv]==4.0.1
python-dotenv==0.19.0
boto3==1.26.0
pytest==6.2.5
pytest-cov==2.12.1
black==21.7b0
flake8==3.9.2
flask-cors==4.0.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.86.0 